Move to separate file and lint
